Move and refine `reorder_question` to QuestionnaireHelper
- Added page refresh to ensure UI updates, e.g., condition creation button visibility.
- Improved validation after drag-and-drop operations.
